Better support webhook payloads containing possible urls (#4241)
Restructure the webhook class contract to better support
payloads that already send all possible clone urls, instead of
guessing a possible list based on host name and repo name.

The existing webhooks (bitbucket, github, gitlab) now implement
GuessUrlWebHookController which takes care of guessing a list of possible
urls as before.

HostedBitBucketController can more cleanly implement the
WebHookController contract because it's not required to provide dummy
values for repo_full_name and repo_host_name.
