+++ server/webapp/WEB-INF/rails/app/controllers/api/web_hooks/bit_bucket_controller.rb
@@ -16,7 +16,7 @@
 
 module Api
   module WebHooks
-    class BitBucketController < WebHookController
+    class BitBucketController < GuessUrlWebHookController
       before_action :verify_content_origin
       before_action :allow_only_push_event
       before_action :allow_only_git_scm
