+++ server/webapp/WEB-INF/rails/app/controllers/api/web_hooks/git_hub_controller.rb
@@ -16,7 +16,7 @@
 
 module Api
   module WebHooks
-    class GitHubController < WebHookController
+    class GitHubController < GuessUrlWebHookController
       before_action :verify_content_origin
       before_action :prempt_ping_call
       before_action :allow_only_push_event
