+++ server/webapp/WEB-INF/rails/app/controllers/api/web_hooks/web_hook_controller.rb
@@ -18,7 +18,7 @@ module Api
   module WebHooks
     class WebHookController < ::Api::ApiController
       def notify
-        Rails.logger.info("[WebHook] Noticed a git push to #{repo_host_name}/#{repo_full_name} on branch #{repo_branch}")
+        Rails.logger.info("[WebHook] Noticed a git push to #{repo_log_name} on branch #{repo_branch}")
 
         if material_update_service.updateGitMaterial(repo_branch, possible_urls)
           render plain: 'OK!', status: :accepted, layout: nil
@@ -28,36 +28,16 @@ module Api
       end
 
       protected
+
       def possible_urls
-        %W(
-          https://#{repo_host_name}/#{repo_full_name}
-          https://#{repo_host_name}/#{repo_full_name}/
-          https://#{repo_host_name}/#{repo_full_name}.git
-          https://#{repo_host_name}/#{repo_full_name}.git/
-          http://#{repo_host_name}/#{repo_full_name}
-          http://#{repo_host_name}/#{repo_full_name}/
-          http://#{repo_host_name}/#{repo_full_name}.git
-          http://#{repo_host_name}/#{repo_full_name}.git/
-          git://#{repo_host_name}/#{repo_full_name}
-          git://#{repo_host_name}/#{repo_full_name}/
-          git://#{repo_host_name}/#{repo_full_name}.git
-          git://#{repo_host_name}/#{repo_full_name}.git/
-          git@#{repo_host_name}:#{repo_full_name}
-          git@#{repo_host_name}:#{repo_full_name}/
-          git@#{repo_host_name}:#{repo_full_name}.git
-          git@#{repo_host_name}:#{repo_full_name}.git/
-        )
+        raise 'Subclass responsibility!'
       end
 
       def repo_branch
         raise 'Subclass responsibility!'
       end
 
-      def repo_host_name
-        raise 'Subclass responsibility!'
-      end
-
-      def repo_full_name
+      def repo_log_name
         raise 'Subclass responsibility!'
       end
